passing tests on RESP2 with Redis 7.2

+++ create.go
@@ -44,7 +44,7 @@ type TextAttribute struct {
 	Sortable       bool
 	UnNormalized   bool
 	Phonetic       string
-	Weight         float32
+	Weight         float64
 	NoStem         bool
 	WithSuffixTrie bool
 	NoIndex        bool
